fix: address #36 review — i128 fee type, negative-fee guard, unused imports
Follow-up to #36 (updatable platform fee and treasury):

- update_fee_bps now takes i128 instead of u32, matching every other
  fee_bps value in the codebase (initialize, storage, ProtocolConfig).
  Switching away from u32 means negative values are no longer ruled
  out by the type system, so the bound check became an explicit
  0..=1_000 range instead of just `> 1_000`.
- FeeUpdated.new_fee_bps updated to i128 to match.
- Documented why update_fee_bps's 1_000 bps cap is intentionally
  tighter than initialize's 0..=10_000 range, rather than leaving it
  looking like a copy-paste of the issue text.
- Removed two unused imports in the admin_updates test module that
  were failing clippy -D warnings.

+++ contracts/campaign-escrow/src/events.rs
@@ -108,7 +108,7 @@ pub struct ContractUpgraded {
 pub struct FeeUpdated {
     #[topic]
     pub admin: Address,
-    pub new_fee_bps: u32,
+    pub new_fee_bps: i128,
 }
 
 #[contractevent]
